+++ src/scripts/actions/observeNodes.ts
@@ -1,5 +1,5 @@
 import { waitElement } from '@1natsu/wait-element'
-import { observeIsVisiblePlayerBar } from '../libs/observeIsVisiblePlayerBar'
+import { observeIsVisiblePlayerBar } from '../controllers/observeIsVisiblePlayerBar'
 
 export const observeNodes = async () => {
   const player = (await waitElement('#movie_player')) as Node
