+++ src/scripts/controllers/observeIsVisiblePlayerBar.ts
