+++ src/scripts/domains/extensionStyle.ts
