+++ src/scripts/presenters/getPlayerBarSize.ts
